requirements.txt
streamlit
torch
torchvision
timm
pillow
numpy
opencv-python
pandas
